Add php website constants "settings"
Addition of the website constants test usage of
unchangable values to be used throughout the site at various places.

Add
- site name, string value
- error reporting, boolean
- site version, number

Along with the error reporting constant, a simply error reporting
array was setup, will grab array errors at various places and
output them at the bottom of the footer when value is TRUE.
Note this included the PHP default error handling with addition
of an basic array used for error handling messages.