Update user logged in check system with HTTP cookie
Add new logged in php constant, removing old `$user_logged_in`
variable, it uses a cookie to determine if the user is logged in
or not. The cookie is set duing the validation of the login.php
script/page - it is set "mid-script" yet avoids 'headers already
sent' error because of output buffering technique in-user.

Also a conditional used to remove the login and register
navigation items based on user logged in constant.