Add session functionalities ---
modify env config file to remove the cookie functionalities that
checked for a logged in user.

modify login to remove previous cookie functionality, and add the
session functionality. After validation, email and a timestamp
are passed to the session superglobal and used in the welcome
page that the user is redirected to.

modify header to remove cookie functionality and replace with
session functionality.

modify welcome to read from session array for email and timestamp
variable.