Update user session and reorganize pages
User session now is now started in the header template file,
instead of a page by page bases. However to check if a logged in
user should be able to access an individual page or not is still
determined by on a page by page bases.

Almost all pages we modified by removing reused header includes
that occurred on each page into the header.php itself, this prevents
repeating of common code, such as including of external scripts. This
also circumvents the page header ('blog header') from being
repeated on each page and is how also housed in the header.php
template file. So each page script will now be focused on information
to be displayed on that page.